[Wip] Add gorilla router package
