+++ examples/gorilla/main.go
@@ -3,19 +3,14 @@ package main
 import (
 	"flag"
 	"log"
-	"net/http"
 	"os"
-	"strings"
-
-	gorilla "github.com/gorilla/mux"
-	"gopkg.in/unrolled/secure.v1"
 
 	"github.com/devopsfaith/krakend/config"
 	"github.com/devopsfaith/krakend/config/viper"
 	"github.com/devopsfaith/krakend/logging"
 	"github.com/devopsfaith/krakend/logging/gologging"
 	"github.com/devopsfaith/krakend/proxy"
-	"github.com/devopsfaith/krakend/router/mux"
+	krakendgorilla "github.com/devopsfaith/krakend/router/gorilla"
 )
 
 func main() {
@@ -41,52 +36,8 @@ func main() {
 		log.Fatal("ERROR:", err.Error())
 	}
 
-	secureMiddleware := secure.New(secure.Options{
-		AllowedHosts:          []string{"127.0.0.1:8080", "example.com", "ssl.example.com"},
-		SSLRedirect:           false,
-		SSLHost:               "ssl.example.com",
-		SSLProxyHeaders:       map[string]string{"X-Forwarded-Proto": "https"},
-		STSSeconds:            315360000,
-		STSIncludeSubdomains:  true,
-		STSPreload:            true,
-		FrameDeny:             true,
-		ContentTypeNosniff:    true,
-		BrowserXssFilter:      true,
-		ContentSecurityPolicy: "default-src 'self'",
-	})
-
-	routerFactory := mux.NewFactory(mux.Config{
-		Engine:         gorillaEngine{gorilla.NewRouter()},
-		ProxyFactory:   customProxyFactory{logger, proxy.DefaultFactory(logger)},
-		Middlewares:    []mux.HandlerMiddleware{secureMiddleware},
-		Logger:         logger,
-		HandlerFactory: mux.CustomEndpointHandler(mux.NewRequestBuilder(gorillaParamsExtractor)),
-		DebugPattern:   "/__debug/{params}",
-	})
-
+	routerFactory := krakendgorilla.DefaultFactory(customProxyFactory{logger, proxy.DefaultFactory(logger)}, logger)
 	routerFactory.New().Run(serviceConfig)
-}
-
-func gorillaParamsExtractor(r *http.Request) map[string]string {
-	params := map[string]string{}
-	for key, value := range gorilla.Vars(r) {
-		params[strings.Title(key)] = value
-	}
-	return params
-}
-
-type gorillaEngine struct {
-	r *gorilla.Router
-}
-
-// Handle implements the mux.Engine interface from the krakend router package
-func (g gorillaEngine) Handle(pattern string, handler http.Handler) {
-	g.r.Handle(pattern, handler)
-}
-
-// ServeHTTP implements the http:Handler interface from the stdlib
-func (g gorillaEngine) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	g.r.ServeHTTP(w, r)
 }
 
 // customProxyFactory adds a logging middleware wrapping the internal factory
